apps/api/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-magic==0.4.27

# AI Providers
openai==1.3.7
anthropic==0.7.8

# Payment
stripe==10.12.0

# HTTP Client
httpx==0.25.2

# Configuration
pydantic>=2.0.1,<3.0.0
pydantic-settings==2.0.3

# File Processing
python-docx==1.1.0
weasyprint==60.2
reportlab==4.0.7
pypdf==3.17.4

# Storage
minio==7.2.0
boto3==1.34.0

# Caching
redis==5.0.1

# Email
fastapi-mail==1.4.1

# Rate Limiting
slowapi==0.1.9

# Logging
structlog==23.2.0
loguru==0.7.2

# Monitoring
sentry-sdk[fastapi]==1.38.0
prometheus-fastapi-instrumentator==6.1.0

# Security & Vulnerability Scanning
safety>=3.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
ruff>=0.14.0
pre-commit==3.6.0
